// frontend/src/api/model/stanceTargetKind.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */

export type StanceTargetKind =
    (typeof StanceTargetKind)[keyof typeof StanceTargetKind];

export const StanceTargetKind = {
    Participant: 'Participant',
    Concept: 'Concept',
    Self: 'Self',
} as const;